% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{create_concordance}
\alias{create_concordance}
\title{Create Concordance}
\usage{
create_concordance(region_from, region_to, id_from = NULL, id_to = NULL)
}
\arguments{
\item{region_from}{A shapefile containing the polygons of the first region type of interest (e.g. SA2).
This should be a sf object, such as that obtained from using st_read() to load
an ESRI shapefile (.shp). Shapefiles for common ABS geographies can be found at
https://www.abs.gov.au/websitedbs/d3310114.nsf/home/digital+boundaries}

\item{region_to}{A shapefile containing the polygons of the second region type of interest (e.g. postcode).
This should be a sf object, such as that obtained from using st_read() to load
an ESRI shapefile (.shp). Shapefiles for common ABS geographies can be found at
https://www.abs.gov.au/websitedbs/d3310114.nsf/home/digital+boundaries}

\item{id_from}{An optional parameter which identifies an ID variable for the first
region (e.g. "sa2_main16"). If supplied along with id_to, these variables will be
used to calculate a clean concordance between the two regions. Otherwise,
only the overlap between regions (in sqm) will be returned, which will require
further manipulation to create a concordance.}

\item{id_to}{See id_from description, but for the regions contained within region_to}
}
\value{
If id_from and id_to are both supplied (and correct), the function will return a
tibble with the ID variables and a concordance from region 1 (region_from), to region 2 (region_to).
If no ID variables are supplied, the function will return a tibble with all
variables in region_from and region_to and the overlap between the two regions in sqm.
}
\description{
This function is designed to create a spatial concordance between two different
sets of regions (e.g. postcode to federal electorate, or postcode to SA2),
based on the spatial information contained in two supplied shapefiles
(one for each region).
}
\details{
This is particularly helpful if you are interested in creating a
concordance between regions, or allocating smaller regions to a larger
region (e.g. postcode to SA3) where there is an imperfect relationship and
publicly available concordances are out of date (e.g. ABS spatial concordances
are limited and have not been updated since 2011, even though regions have changed).
}
\examples{
# Generate a concordance between Australian Postcodes and SA3 regions
(these shapefiles are available for download from here: https://www.abs.gov.au/websitedbs/d3310114.nsf/home/digital+boundaries)

sa3_shp <- st_read("SA3_2021_AUST_GDA2020.shp")

poa_shp <- st_read("POA_2021_AUST_GDA2020.shp")

poa_to_sa3 <-
  create_concordance(
    region_from = poa_shp,
    region_to = sa3_shp,
    id_from = "POA_CODE21",
    id_to = "SA3_NAME21"
  )

}
